# server/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Kekkon API
  description: API untuk aplikasi undangan pernikahan digital Kekkon
  version: 1.0.0
  contact:
    name: Kekkon Support
    url: https://kekkon.hafizhrf.me
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://kekkon.hafizhrf.me
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
    Invitation:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        template_id:
          type: string
        bride_name:
          type: string
        groom_name:
          type: string
        wedding_date:
          type: string
          format: date
        akad_time:
          type: string
        akad_venue:
          type: string
        akad_address:
          type: string
        reception_time:
          type: string
        reception_venue:
          type: string
        reception_address:
          type: string
        primary_color:
          type: string
        secondary_color:
          type: string
        gallery_images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              caption:
                type: string
        gift_bank_accounts:
          type: array
          items:
            type: object
            properties:
              bank:
                type: string
              number:
                type: string
              name:
                type: string
...
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        category:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  minLength: 2
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
...
  /api/invitations:
    get:
      tags: [Invitations]
      summary: Get all invitations for current user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of invitations
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invitation'

    post:
      tags: [Invitations]
      summary: Create new invitation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [template_id, bride_name, groom_name]
              properties:
                template_id:
                  type: string
                bride_name:
                  type: string
                groom_name:
                  type: string
                wedding_date:
                  type: string
                  format: date
                akad_time:
                  type: string
                akad_venue:
                  type: string
                akad_address:
                  type: string
                reception_time:
                  type: string
                reception_venue:
                  type: string
                reception_address:
                  type: string
                primary_color:
                  type: string
                secondary_color:
                  type: string
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
  /api/invitations/{id}:
    get:
      tags: [Invitations]
      summary: Get invitation by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invitation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '404':
          description: Invitation not found

    put:
      tags: [Invitations]
      summary: Update invitation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invitation'
      responses:
        '200':
          description: Invitation updated

    delete:
      tags: [Invitations]
      summary: Delete invitation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invitation deleted
...
  /api/templates/{id}:
    get:
      tags: [Templates]
      summary: Get template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'

  # Public
  /api/public/{slug}:
    get:
      tags: [Public]
      summary: Get public invitation
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
          description: Guest name for personalization
      responses:
        '200':
          description: Public invitation data
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitation:
                    $ref: '#/components/schemas/Invitation'
                  guestName:
                    type: string
...
  /api/admin/invitations:
    get:
      tags: [Admin]
      summary: Get all invitations (admin)
      security:
        - AdminAuth: []
      responses:
        '200':
          description: List of all invitations
...